Fix first big chunk of import paths. Remove * imports. Move fits_are_identical to shared utilities.
