+++ python/a_download_decals/get_catalogs/selection_cuts_test.py
@@ -1,7 +1,7 @@
 
 import pytest
 
-from get_catalogs.selection_cuts import *
+from a_download_decals.get_catalogs.selection_cuts import *
 from astropy.table import Table
