+++ python/a_download_decals/get_images/image_utils_test.py
@@ -3,7 +3,7 @@ import pytest
 from astropy.io import fits
 
 
-from get_images.image_utils import dr2_style_rgb, decals_internal_rgb, lupton_rgb
+from a_download_decals.get_images.image_utils import dr2_style_rgb, decals_internal_rgb, lupton_rgb
 
 TEST_EXAMPLES_DIR = 'python/test_examples'
