+++ python/b_to_zooniverse/make_calibration_images/get_calibration_images.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 from astropy.io import fits
 from tqdm import tqdm
 
-from get_images import image_utils
+from a_download_decals.get_images import image_utils
 from a_download_decals.get_images.download_images_threaded import get_loc
